+++ packages/next/src/client/components/react-dev-overlay/pages/hot-reloader-client.ts
@@ -35,18 +35,18 @@ import {
   onBeforeRefresh,
   onRefresh,
   onVersionInfo,
-} from '../../components/react-dev-overlay/pages/client'
+} from './client'
 import stripAnsi from 'next/dist/compiled/strip-ansi'
 import { addMessageListener, sendMessage } from './websocket'
-import formatWebpackMessages from './format-webpack-messages'
-import { HMR_ACTIONS_SENT_TO_BROWSER } from '../../../server/dev/hot-reloader-types'
+import formatWebpackMessages from '../internal/helpers/format-webpack-messages'
+import { HMR_ACTIONS_SENT_TO_BROWSER } from '../../../../server/dev/hot-reloader-types'
 import type {
   HMR_ACTION_TYPES,
   TurbopackMsgToBrowser,
-} from '../../../server/dev/hot-reloader-types'
-import { extractModulesFromTurbopackMessage } from '../../../server/dev/extract-modules-from-turbopack-message'
-import { RuntimeErrorHandler } from '../../components/react-dev-overlay/internal/helpers/runtime-error-handler'
-import { REACT_REFRESH_FULL_RELOAD_FROM_ERROR } from './messages'
+} from '../../../../server/dev/hot-reloader-types'
+import { extractModulesFromTurbopackMessage } from '../../../../server/dev/extract-modules-from-turbopack-message'
+import { RuntimeErrorHandler } from '../internal/helpers/runtime-error-handler'
+import { REACT_REFRESH_FULL_RELOAD_FROM_ERROR } from '../shared'
 // This alternative WebpackDevServer combines the functionality of:
 // https://github.com/webpack/webpack-dev-server/blob/webpack-1/client/index.js
 // https://github.com/webpack/webpack/blob/webpack-1/hot/dev-server.js
