+++ packages/next/src/client/components/react-dev-overlay/shared.ts
@@ -161,3 +161,6 @@ export function useErrorOverlayReducer() {
     }
   }, INITIAL_OVERLAY_STATE)
 }
+
+export const REACT_REFRESH_FULL_RELOAD_FROM_ERROR =
+  '[Fast Refresh] performing full reload because your application had an unrecoverable error'
