+++ packages/next/src/client/dev/amp-dev.ts
@@ -1,7 +1,10 @@
 /* globals __webpack_hash__ */
 import { displayContent } from './fouc'
 import initOnDemandEntries from './on-demand-entries-client'
-import { addMessageListener, connectHMR } from './error-overlay/websocket'
+import {
+  addMessageListener,
+  connectHMR,
+} from '../components/react-dev-overlay/pages/websocket'
 import { HMR_ACTIONS_SENT_TO_BROWSER } from '../../server/dev/hot-reloader-types'
 
 declare global {
