--- packages/next/src/client/dev/error-overlay/messages.ts
@@ -1,2 +1,0 @@
-export const REACT_REFRESH_FULL_RELOAD_FROM_ERROR =
-  '[Fast Refresh] performing full reload because your application had an unrecoverable error'
