+++ packages/next/src/client/tracing/report-to-socket.ts
@@ -1,4 +1,4 @@
-import { sendMessage } from '../dev/error-overlay/websocket'
+import { sendMessage } from '../components/react-dev-overlay/pages/websocket'
 import type { Span } from './tracer'
 
 export default function reportToSocket(span: Span) {
